!fixup rename LazyTransform to lazy_transform
