+++ lib/crypto.js
@@ -20,7 +20,7 @@ const constants = require('constants');
 const stream = require('stream');
 const util = require('util');
 const internalUtil = require('internal/util');
-const LazyTransform = require('internal/streams/LazyTransform');
+const LazyTransform = require('internal/streams/lazy_transform');
 
 const DH_GENERATOR = 2;
