+++ lib/internal/streams/lazy_transform.js
