+++ node.gyp
@@ -73,7 +73,7 @@
       'lib/internal/socket_list.js',
       'lib/internal/repl.js',
       'lib/internal/util.js',
-      'lib/internal/streams/LazyTransform.js',
+      'lib/internal/streams/lazy_transform.js',
     ],
   },
